requirements.txt
Django==5.0.9
djangorestframework==3.15.2
django-cors-headers==4.4.0
psycopg2-binary==2.9.9
Pillow==10.4.0
requests==2.32.3
celery==5.4.0
redis==5.0.8
python-decouple==3.8
django-filter==24.3
drf-spectacular==0.27.2